we lifted state up and used props to forward them to the UserInput Component
