+++ src/App.jsx
@@ -1,11 +1,29 @@
+import { useState } from "react";
+
 import Header from "./components/Header";
 import UserInput from "./components/UserInput";
 
 function App() {
+  const [userInput, setUserInput] = useState({
+    initialInvestment: 10000,
+    annualInvestment: 1200,
+    expectedReturn: 6,
+    duration: 10,
+  });
+
+  function handleChange(inputIdentifier, newValue) {
+    setUserInput((prevUserInput) => {
+      return {
+        ...prevUserInput,
+        [inputIdentifier]: newValue,
+      };
+    });
+  }
+
   return (
     <>
       <Header />
-      <UserInput />
+      <UserInput userInput={userInput} onChange={handleChange} />
     </>
   );
 }
